+++ vistle/module/general/CutGeometry/CutGeometry.cpp
@@ -150,7 +150,7 @@ class PlaneClip {
          outIdxCorner[0] = 0;
          outIdxCoord[0] = numCoordsIn;
 #pragma omp parallel for schedule (dynamic)
-         for (ssize_t elem=0; elem<numElem; ++elem) {
+         for (ssize_t elem=0; elem<ssize_t(numElem); ++elem) {
 
             Index numCorner=0, numCoord=0;
             processTriangle(elem, numCorner, numCoord, true);
@@ -183,7 +183,7 @@ class PlaneClip {
          }
 
 #pragma omp parallel for schedule (dynamic)
-         for (ssize_t elem = 0; elem<numElem; ++elem) {
+         for (ssize_t elem = 0; elem<ssize_t(numElem); ++elem) {
             processTriangle(elem, outIdxCorner[elem], outIdxCoord[elem], false);
          }
          //std::cerr << "CuttingSurface: << " << m_outData->x().size() << " vert, " << m_outData->x().size() << " data elements" << std::endl;
@@ -198,7 +198,7 @@ class PlaneClip {
          outIdxCorner[0] = 0;
          outIdxCoord[0] = numCoordsIn;
 #pragma omp parallel for schedule (dynamic)
-         for (ssize_t elem=0; elem<numElem; ++elem) {
+         for (ssize_t elem=0; elem<ssize_t(numElem); ++elem) {
 
             Index numPoly=0, numCorner=0, numCoord=0;
             processPolygon(elem, numPoly, numCorner, numCoord, true);
@@ -233,7 +233,7 @@ class PlaneClip {
          }
 
 #pragma omp parallel for schedule (dynamic)
-         for (ssize_t elem = 0; elem<numElem; ++elem) {
+         for (ssize_t elem = 0; elem<ssize_t(numElem); ++elem) {
             processPolygon(elem, outIdxPoly[elem], outIdxCorner[elem], outIdxCoord[elem], false);
          }
 
@@ -257,7 +257,7 @@ class PlaneClip {
       m_vertexMap.resize(nCoord);
       auto vertexMap = m_vertexMap.data();
 #pragma omp parallel for
-      for (ssize_t i=0; i<nCoord; ++i) {
+      for (ssize_t i=0; i<ssize_t(nCoord); ++i) {
          const Vector p(x[i], y[i], z[i]);
          vertexMap[i] = m_decider(i) > 0 ? 1 : 0;
       }
@@ -276,7 +276,7 @@ class PlaneClip {
          out_y = m_outCoords->y().data();
          out_z = m_outCoords->z().data();
 #pragma omp parallel for schedule(dynamic)
-         for (ssize_t i=0; i<nCoord; ++i) {
+         for (ssize_t i=0; i<ssize_t(nCoord); ++i) {
             Index idx = vertexMap[i];
             assert(idx >= 0);
             if (idx > 0) {
